@@ -4,35 +4,37 @@ publish_to: "none"
 version: 1.4.1+18
 
 environment:
-  sdk: ">=2.7.0 <3.0.0"
+  sdk: ">=2.12.0 <3.0.0"
 
 dependencies:
-  cached_network_image: ^2.5.1
-  dio: ^3.0.10
-  fluentui_system_icons: ^1.1.109
+  cached_network_image: ^3.0.0-nullsafety
+  dio: ^4.0.0-prev2
+  fluentui_system_icons: ^1.1.110
   flutter:
     sdk: flutter
   flutter_secure_storage: ^4.0.0
   flutter_speed_dial:
     git:
       url: https://github.com/darioielardi/flutter_speed_dial
-      ref: 282dd8956404944972c5a7eac51d52690c7582f5
+      ref: new
   flutter_widget_from_html_core: ^0.6.0-rc
-  hooks_riverpod: ^0.12.4
+  hooks_riverpod: ^0.13.1
   html_unescape: ^2.0.0-nullsafety
   jiffy: ^4.0.0
   share: ^2.0.0
-  shared_preferences: ^2.0.4
+  shared_preferences: ^2.0.5
   shimmer: ^2.0.0-nullsafety
-  uni_links: ^0.4.0
+  string_validator: ^0.3.0
+  uni_links:
+    git:
+      url: https://github.com/objectivecosta/uni_links_simplified
   url_launcher: ^6.0.0
-  validators: ^2.0.1
 
 dev_dependencies:
   build_runner: ^1.11.5
-  flutter_launcher_icons: ^0.8.1
-  freezed: ^0.12.7
-  json_serializable: ^3.5.1
+  flutter_launcher_icons: ^0.9.0
+  freezed: ^0.14.1
+  json_serializable: ^4.0.2
 
 flutter:
   uses-material-design: true
