@@ -5,11 +5,11 @@ import 'package:glider/utils/app_bar_util.dart';
 import 'package:glider/widgets/reply/reply_body.dart';
 
 class ReplyPage extends HookWidget {
-  const ReplyPage({Key key, @required this.parent, this.root})
+  const ReplyPage({Key? key, required this.parent, this.root})
       : super(key: key);
 
   final Item parent;
-  final Item root;
+  final Item? root;
 
   @override
   Widget build(BuildContext context) {
