@@ -1,7 +1,7 @@
 class ServiceException implements Exception {
   ServiceException([this.message]);
 
-  final String message;
+  final String? message;
 
   @override
   String toString() {
