@@ -8,7 +8,7 @@ import 'package:glider/widgets/common/loading.dart';
 import 'package:hooks_riverpod/hooks_riverpod.dart';
 
 class AccountBody extends HookWidget {
-  const AccountBody({Key key}) : super(key: key);
+  const AccountBody({Key? key}) : super(key: key);
 
   @override
   Widget build(BuildContext context) {
