@@ -1,7 +1,7 @@
 import 'package:flutter/material.dart';
 
 class Loading extends StatelessWidget {
-  const Loading({Key key}) : super(key: key);
+  const Loading({Key? key}) : super(key: key);
 
   @override
   Widget build(BuildContext context) {
