@@ -3,17 +3,17 @@ import 'package:glider/utils/text_style_extension.dart';
 
 class MetadataItem extends StatelessWidget {
   const MetadataItem(
-      {Key key, @required this.icon, this.text, this.highlight = false})
+      {Key? key, required this.icon, this.text, this.highlight = false})
       : super(key: key);
 
   final IconData icon;
-  final String text;
+  final String? text;
   final bool highlight;
 
   @override
   Widget build(BuildContext context) {
     final TextTheme textTheme = Theme.of(context).textTheme;
-    final Color color =
+    final Color? color =
         highlight ? Theme.of(context).colorScheme.primary : null;
 
     return Padding(
@@ -22,15 +22,15 @@ class MetadataItem extends StatelessWidget {
         children: <Widget>[
           Icon(
             icon,
-            size: textTheme.bodyText2.scaledFontSize(context),
+            size: textTheme.bodyText2?.scaledFontSize(context),
             color: color,
           ),
           if (text != null) ...<Widget>[
             const SizedBox(width: 4),
             Text(
-              text,
-              style: textTheme.bodyText2.copyWith(
-                fontSize: textTheme.caption.fontSize,
+              text!,
+              style: textTheme.bodyText2?.copyWith(
+                fontSize: textTheme.caption?.scaledFontSize(context),
                 color: color,
               ),
             ),
