@@ -3,36 +3,36 @@ import 'package:glider/utils/animation_util.dart';
 
 class SmoothAnimatedSwitcher extends StatelessWidget {
   const SmoothAnimatedSwitcher({
-    Key key,
-    @required this.condition,
-    @required this.child,
+    Key? key,
+    required this.condition,
+    required this.child,
     this.duration,
   })  : transitionBuilder = AnimationUtil.fadeTransitionBuilder,
         super(key: key);
 
   const SmoothAnimatedSwitcher.vertical({
-    Key key,
-    @required this.condition,
-    @required this.child,
+    Key? key,
+    required this.condition,
+    required this.child,
     this.duration,
   })  : transitionBuilder = AnimationUtil.verticalFadeTransitionBuilder,
         super(key: key);
 
   const SmoothAnimatedSwitcher.horizontal({
-    Key key,
-    @required this.condition,
-    @required this.child,
+    Key? key,
+    required this.condition,
+    required this.child,
     this.duration,
   })  : transitionBuilder = AnimationUtil.horizontalFadeTransitionBuilder,
         super(key: key);
 
   final bool condition;
   final Widget child;
-  final Duration duration;
+  final Duration? duration;
   final AnimatedSwitcherTransitionBuilder transitionBuilder;
 
   static Widget _defaultLayoutBuilder(
-      Widget currentChild, List<Widget> previousChildren) {
+      Widget? currentChild, List<Widget> previousChildren) {
     return Stack(
       children: <Widget>[
         ...previousChildren,
