@@ -7,21 +7,21 @@ import 'package:hooks_riverpod/hooks_riverpod.dart';
 
 class ItemTile extends HookWidget {
   const ItemTile({
-    Key key,
-    @required this.id,
-    this.ancestors,
+    Key? key,
+    required this.id,
+    this.ancestors = const <int>[],
     this.root,
     this.onTap,
     this.dense = false,
     this.interactive = false,
     this.fadeable = false,
-    @required this.loading,
+    required this.loading,
   }) : super(key: key);
 
   final int id;
   final Iterable<int> ancestors;
-  final Item root;
-  final void Function(BuildContext) onTap;
+  final Item? root;
+  final void Function(BuildContext)? onTap;
   final bool dense;
   final bool interactive;
   final bool fadeable;
@@ -29,22 +29,24 @@ class ItemTile extends HookWidget {
 
   @override
   Widget build(BuildContext context) {
-    final Item item = useProvider(itemCacheStateProvider(id)).state ??
-        useProvider(itemProvider(id)).data?.value;
+    final Item? cachedItem = useProvider(itemCacheStateProvider(id)).state;
+    final AsyncData<Item?>? itemData = cachedItem != null
+        ? AsyncData<Item?>(cachedItem)
+        : useProvider(itemProvider(id)).data;
 
-    if (item == null) {
+    if (itemData == null) {
       return loading();
     }
 
-    if (item.time == null) {
+    if (itemData.value?.time == null) {
       return const SizedBox.shrink();
     }
 
     return ItemTileData(
-      item.copyWith(ancestors: ancestors),
+      itemData.value!.copyWith(ancestors: ancestors),
       key: ValueKey<int>(id),
       root: root,
-      onTap: () => onTap(context),
+      onTap: () => onTap?.call(context),
       dense: dense,
       interactive: interactive,
       fadeable: fadeable,
