@@ -9,16 +9,16 @@ import 'package:hooks_riverpod/hooks_riverpod.dart';
 class ItemTileContentPollOption extends HookWidget {
   const ItemTileContentPollOption(
     this.item, {
-    Key key,
+    Key? key,
     this.root,
     this.interactive = false,
-    @required this.vote,
+    required this.vote,
   }) : super(key: key);
 
   final Item item;
-  final Item root;
+  final Item? root;
   final bool interactive;
-  final Future<void> Function(BuildContext, {bool up}) vote;
+  final Future<void> Function(BuildContext, {required bool up}) vote;
 
   @override
   Widget build(BuildContext context) {
@@ -29,10 +29,16 @@ class ItemTileContentPollOption extends HookWidget {
             data: (bool upvoted) => upvoted,
             orElse: () => false,
           ),
-          onChanged: interactive ? (bool up) => vote(context, up: up) : null,
+          onChanged:
+              interactive ? (bool? up) => vote(context, up: up == true) : null,
           materialTapTargetSize: MaterialTapTargetSize.shrinkWrap,
         ),
-        Expanded(child: ItemTileText(item)),
+        if (item.text != null)
+          Expanded(
+            child: ItemTileText(item),
+          )
+        else
+          const Spacer(),
         ItemTileMetadata(item, root: root),
       ],
     );
