@@ -2,7 +2,7 @@ import 'package:flutter/material.dart';
 import 'package:flutter_hooks/flutter_hooks.dart';
 
 class ItemTilePreview extends HookWidget {
-  const ItemTilePreview({Key key}) : super(key: key);
+  const ItemTilePreview({Key? key}) : super(key: key);
 
   @override
   Widget build(BuildContext context) {
