@@ -4,7 +4,7 @@ import 'package:glider/models/item.dart';
 import 'package:glider/widgets/common/decorated_html.dart';
 
 class ItemTileText extends HookWidget {
-  const ItemTileText(this.item, {Key key}) : super(key: key);
+  const ItemTileText(this.item, {Key? key}) : super(key: key);
 
   final Item item;
 
@@ -12,7 +12,7 @@ class ItemTileText extends HookWidget {
   Widget build(BuildContext context) {
     return Hero(
       tag: 'item_text_${item.id}',
-      child: DecoratedHtml(item.text),
+      child: DecoratedHtml(item.text!),
     );
   }
 }
