@@ -6,7 +6,7 @@ import 'package:glider/widgets/common/metadata_item.dart';
 import 'package:glider/widgets/users/user_bottom_sheet.dart';
 
 class UserTileData extends StatelessWidget {
-  const UserTileData(this.user, {Key key}) : super(key: key);
+  const UserTileData(this.user, {Key? key}) : super(key: key);
 
   final User user;
 
@@ -23,27 +23,23 @@ class UserTileData extends StatelessWidget {
           children: <Widget>[
             Row(
               children: <Widget>[
-                if (user.karma != null)
-                  MetadataItem(
-                    icon: FluentIcons.arrow_up_24_regular,
-                    text: user.karma.toString(),
+                MetadataItem(
+                  icon: FluentIcons.arrow_up_24_regular,
+                  text: user.karma.toString(),
+                ),
+                MetadataItem(
+                  icon: FluentIcons.comment_24_regular,
+                  text: user.submitted.length.toString(),
+                ),
+                Padding(
+                  padding: const EdgeInsets.symmetric(vertical: 1),
+                  child: Text(
+                    user.id,
+                    style: textTheme.caption?.copyWith(
+                        color: Theme.of(context).colorScheme.primary),
                   ),
-                if (user.submitted != null)
-                  MetadataItem(
-                    icon: FluentIcons.comment_24_regular,
-                    text: user.submitted.length.toString(),
-                  ),
-                if (user.id != null) ...<Widget>[
-                  Padding(
-                    padding: const EdgeInsets.symmetric(vertical: 1),
-                    child: Text(
-                      user.id,
-                      style: textTheme.caption.copyWith(
-                          color: Theme.of(context).colorScheme.primary),
-                    ),
-                  ),
-                  const SizedBox(width: 8),
-                ],
+                ),
+                const SizedBox(width: 8),
                 const Spacer(),
                 Text(
                   'since ${user.createdDate}',
@@ -53,7 +49,7 @@ class UserTileData extends StatelessWidget {
             ),
             if (user.about != null) ...<Widget>[
               const SizedBox(height: 12),
-              DecoratedHtml(user.about),
+              DecoratedHtml(user.about!),
             ],
           ],
         ),
