@@ -7,24 +7,24 @@ part 'item.freezed.dart';
 part 'item.g.dart';
 
 @freezed
-abstract class Item with _$Item {
+class Item with _$Item {
   factory Item({
-    int id,
-    bool deleted,
-    ItemType type,
-    String by,
-    int time,
-    String text,
-    bool dead,
-    int parent,
-    int poll,
-    Iterable<int> kids,
-    String url,
-    int score,
-    String title,
-    Iterable<int> parts,
-    int descendants,
-    Iterable<int> ancestors,
+    required int id,
+    bool? deleted,
+    ItemType? type,
+    String? by,
+    int? time,
+    String? text,
+    bool? dead,
+    int? parent,
+    int? poll,
+    @Default(<int>[]) Iterable<int> kids,
+    String? url,
+    int? score,
+    String? title,
+    @Default(<int>[]) Iterable<int> parts,
+    int? descendants,
+    @Default(<int>[]) Iterable<int> ancestors,
   }) = _Item;
 
   factory Item.fromJson(Map<String, dynamic> json) => _$ItemFromJson(json);
@@ -35,33 +35,34 @@ extension ItemExtension on Item {
   static final RegExp _pdfRegExp = RegExp(r'\s+\[pdf\]$');
   static final RegExp _yearRegExp = RegExp(r'\s+\((\d+)\)');
 
-  String get taglessTitle => title
+  String? get taglessTitle => title
       ?.replaceAll(_videoRegExp, '')
-      ?.replaceAll(_pdfRegExp, '')
-      ?.replaceAll(_yearRegExp, '');
+      .replaceAll(_pdfRegExp, '')
+      .replaceAll(_yearRegExp, '');
 
-  String get urlHost => Uri.parse(url)?.host;
+  String? get urlHost {
+    final String? url = this.url;
+    return url != null ? Uri.parse(url).host : null;
+  }
 
-  String get timeAgo => Jiffy.unix(time).fromNow();
+  String? get timeAgo => time != null ? Jiffy.unix(time!).fromNow() : null;
 
-  String get thumbnailUrl =>
+  String? get thumbnailUrl =>
       localOnly ? null : 'https://drcs9k8uelb9s.cloudfront.net/$id.png';
 
-  bool get localOnly => id != null && id < 0;
+  bool get localOnly => id < 0;
 
-  bool get hasVideo => title != null ? _videoRegExp.hasMatch(title) : null;
+  bool get hasVideo => title != null && _videoRegExp.hasMatch(title!);
 
-  bool get hasPdf => title != null ? _pdfRegExp.hasMatch(title) : null;
+  bool get hasPdf => title != null && _pdfRegExp.hasMatch(title!);
 
-  bool get hasOriginalYear =>
-      title != null ? _yearRegExp.hasMatch(title) : null;
+  bool get hasOriginalYear => title != null && _yearRegExp.hasMatch(title!);
 
-  String get originalYear =>
-      title != null ? _yearRegExp.firstMatch(title)?.group(1) : null;
+  String? get originalYear =>
+      title != null ? _yearRegExp.firstMatch(title!)?.group(1) : null;
 
-  Item addKid(int kidId) =>
-      copyWith(kids: <int>[kidId, if (kids != null) ...kids]);
+  Item addKid(int kidId) => copyWith(kids: <int>[kidId, ...kids]);
 
   Item incrementDescendants() =>
-      copyWith(descendants: descendants != null ? descendants + 1 : null);
+      copyWith(descendants: descendants != null ? descendants! + 1 : null);
 }
