@@ -1,6 +1,7 @@
 import 'package:flutter/widgets.dart';
 
 extension TextStyleExtension on TextStyle {
-  double scaledFontSize(BuildContext context) =>
-      fontSize * MediaQuery.of(context).textScaleFactor;
+  double? scaledFontSize(BuildContext context) => fontSize != null
+      ? fontSize! * MediaQuery.of(context).textScaleFactor
+      : null;
 }
