@@ -1,9 +1,9 @@
-import 'package:validators/validators.dart';
+import 'package:string_validator/string_validator.dart';
 
 class Validators {
   Validators._();
 
-  static String notEmpty(String value) {
+  static String? notEmpty(String? value) {
     if (value == null || value.isEmpty) {
       return 'This field cannot be empty.';
     }
@@ -11,7 +11,7 @@ class Validators {
     return null;
   }
 
-  static String maxLength(String value, int maxLength) {
+  static String? maxLength(String? value, int maxLength) {
     if (value != null && value.length > maxLength) {
       return 'Value must have a length less than or equal to $maxLength.';
     }
@@ -19,8 +19,9 @@ class Validators {
     return null;
   }
 
-  static String url(String value) {
-    if (!isURL(value, requireProtocol: true)) {
+  static String? url(String? value) {
+    if (value != null &&
+        !isURL(value, <String, Object>{'requireProtocol': true})) {
       return 'This field requires a valid URL address.';
     }
 
